+++ app/scripts/load_script_generator.js
@@ -23,6 +23,28 @@ window.LI = window.LI || {};
     var BATCH_THRESHOLD = 3000;
     var PAGE_MATCH_THRESHOLD = 1000;
 
+    var utf8ArrayToStr = function(array) {
+      // avoid Maximum call stack size exceeded
+      ///https://stackoverflow.com/questions/8936984/uint8array-to-string-in-javascript
+      //https://bugs.webkit.org/show_bug.cgi?id=80797
+      var result;
+      try {
+        result = String.fromCharCode.apply(null, new Uint8Array(array));
+      } catch (e) {
+        result = "";
+      }
+      return result;
+    };
+
+    var escapeUrl = function(input) {
+      if (input) {
+
+        // \[^"] --> \\[^"] (escape backslashes to work in Lua)
+        input = input.replace(/\\([^"])/g, '\\\\$1');
+      }
+      return input;
+    };
+
     var escapeContent = function(input) {
       if (input) {
 
@@ -30,7 +52,7 @@ window.LI = window.LI || {};
         input = input.replace(/"/g, '\\\"');
 
         //  This replace fixes content which contains a doublequote string
-        //  \"  - (first replace) -> \\" - (second replace to work in LUA) -> \\\"
+        //  \"  - (first replace) -> \\" - (second replace to work in Lua) -> \\\"
         input = input.replace(/\\\\"/g, '\\\\\\\"');
         input = input.replace(/[\r\n]/g, "");
       }
@@ -58,6 +80,7 @@ window.LI = window.LI || {};
         // Build IR from requests.
         transactions.forEach(function(transaction) {
             var time = transaction.timeStamp;
+
 
             // Determine if this transaction represents a new page.
             if (self._isTransactionNewPage(transaction)) {
@@ -210,10 +233,30 @@ window.LI = window.LI || {};
 
     LI.LoadScriptGenerator.prototype._compileIRToLua = function(ir) {
         var self = this,
+            isFirstBatch = true,
+            isFirstRequest = true,
             script = '';
+
+        script += '-- The chrome extension has recorded all your HTTP traffic to generate this user scenario script.\n';
+        script += '-- You can use this user scenario in your load tests to simulate your user traffic.\n';
+        script +=  '\n\n';
+        script += '-- You will likely have to modify this script to:\n';
+        script += '--   1) Rename the page name `http.page_start/http.page_end` to help during result analysis.\n';
+        script += '--   2) Setup more accurate sleep time `client.sleep`. Sleep time is important, because real users do go idle on pages, depending on content.\n';
+        script += '--   3) Extracting and correlating CSRF tokens or dynamic values.\n';
+        script += '--   4) Adding custom metrics, logs, custom logic...\n';
+        script += '\n\n';
+        script += '-- A script validation launches 1 Virtual User to run your user scenario; it shows logs, warnings and errors.\n';
+        script += '\n\n';
+
         ir.forEach(function(node) {
             if ('batch' === node[0]) {
-                script += "http.request_batch({\n";
+
+                if (isFirstBatch) {
+                  script += "responses = http.request_batch({\n";
+                } else {
+                  script += "http.request_batch({\n";
+                }
                 var requests = [];
                 node[1].forEach(function(transaction) {
                     var method = transaction.method.toUpperCase(),
@@ -242,7 +285,7 @@ window.LI = window.LI || {};
                                               'image/png',
                                               'image/tiff',
                                               'multipart/form-data'],
-                        requestIR = [method, transaction.url];
+                        requestIR = [method, escapeUrl(transaction.url)];
 
                     // Add X-headers.
                     if (transaction.requestHeaders) {
@@ -275,27 +318,35 @@ window.LI = window.LI || {};
                                              body[0] && body[0].bytes &&
                                              body[0].bytes instanceof ArrayBuffer);
 
+                        if (!isArrayBuffer && $.inArray(method, bodyMethods) && '' !== contentType && $.inArray(contentType, base64ContentTypes)) {
+                          shouldBase64Body = true;
+                        } else {
 
-                        if (!isArrayBuffer) {
-                          if ($.inArray(method, bodyMethods) &&
-                              '' !== contentType &&
-                              $.inArray(contentType, base64ContentTypes)) {
-                              shouldBase64Body = true;
-                          } else if (/[\x00-\x08\x0E-\x1F\x80-\xFF]/.test(body)) { // Look for binary data.
-                              shouldBase64Body = true;
-                          }
+                          //http://stackoverflow.com/questions/1677644/detect-non-printable-characters-in-javascript
+                          //
+                          // Non printable characters will break the script output.
+                          //
+                          // Non printable characters must be escaped. Because they usually are binary data,
+                          // the data is encoded as base64
+
+                          var text = isArrayBuffer ? utf8ArrayToStr(body[0].bytes) : body;
+                          shouldBase64Body = /[\x00-\x08\x0E-\x1F\x80-\xFF]/.test(text);
                         }
 
 
+
                         if (shouldBase64Body) {
-                            requestIR.push(['data', '"' + btoa(body[0].bytes) + '"']);
+
+                            var bodyContent = isArrayBuffer ? utf8ArrayToStr(body[0].bytes) : body[0].bytes;
+
+                            requestIR.push(['data', '"' + btoa(bodyContent) + '"']);
                             requestIR.push(['base64_encoded_body', 'true']);
                         } else {
                             if ($.isPlainObject(body)) {
                                 requestIR.push(['data', '"' + self._convertFormDataToBodyData(body) + '"']);
                             } else {
                                 if (body && body[0] && body[0].bytes) {
-                                  var bodyAsString = String.fromCharCode.apply(null, new Uint8Array(body[0].bytes));
+                                  var bodyAsString = utf8ArrayToStr(body[0].bytes);
                                   requestIR.push(['data', '"' + escapeContent(bodyAsString) + '"']);
                                 }
                             }
@@ -304,13 +355,46 @@ window.LI = window.LI || {};
                         requestIR.push(['headers', headers]);
                     }
 
+
                     // Add compression handling.
                     requestIR.push(['auto_decompress', 'true']);
+
+                    if (isFirstRequest) {
+                      requestIR.push(['response_body_bytes', 1024]);
+                      isFirstRequest = false;
+                    }
 
                     // Compile request IR into a Lua string.
                     requests.push("\t{" + self._compileRequestIRToLua(requestIR) + "}");
                 });
-                script += requests.join(",\n") + "\n})\n\n";
+
+
+                if (isFirstBatch) {
+
+                  if (requests.length) {
+                    var firstRequests = requests.splice(0, 1);
+                    script += firstRequests.join(",\n") + "\n})\n\n";
+
+                    script += '\n\n\n';
+                    script += '-- Use `response_body_bytes` parameter to access response body or use `http.set_option()`\n';
+                    script += '-- ** Feel free to remove the line of code below\n';
+                    script += 'log.debug(\"FirstRequestBody: \"..responses[1].body)\n';
+                    script += '\n';
+                    script += '-- Report any specific data with `result.custom_metric`\n';
+                    script += '-- ** Feel free to remove the line of code below\n';
+                    script += 'result.custom_metric("TTFB-FirstRequest", responses[1].time_to_first_byte)\n';
+                    script += '\n\n\n';
+
+                    if (requests.length) {
+                      script += requests.join(",\n") + "\n})\n\n";
+                    }
+                  }
+
+                  isFirstBatch = false;
+
+                } else {
+                  script += requests.join(",\n") + "\n})\n\n";
+                }
             } else if ('pageStart' === node[0]) {
                 script += 'http.page_start("' + node[1] + '")\n';
             } else if ('pageEnd' === node[0]) {