+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+### 1.0.15-20180110
+
+- [fix] Escape backslashes in URLs
+
+### 1.0.14-20171118
+
+- [fix] Refactor the fix with `out of range` body size
+
+### 1.0.13-20171116
+
+- [fix] Fix `tabId` filter was not working correctly and requests were not recorded
+
+### 1.0.12-20171113
+
+- [fix] Fix issue with `out of range` body size
+
+### 1.0.11-20160922
+
+- [feat] Add onboarding code and comments to the generated user scenario
+
+### 1.0.10-20160609
+
+- [fix] Base64 encode non-printable characters when body isArrayBuffer data and Content-Type header is not specified
+
+### 1.0.9-20160329
+- [fix] Allow to cancel recording of another browser tab
 
 ### 1.0.8-20151130
 - [fix] Strings with backslash-escaped double quotes are